list tags component complete
