+++ src/components/ListTags/ListTags.jsx
@@ -4,6 +4,8 @@ import Popover from '@mui/material/Popover';
 import Typography from '@mui/material/Typography';
 import Button from '@mui/material/Button';
 
+import Tag from '../Tag/Tag';
+
 export default function ListTags(props){
     let defaultTags = [
         {id: 0, name: 'tag0', description: 'first tag'},
@@ -18,41 +20,15 @@ export default function ListTags(props){
     let tagsArray = props.tags || defaultTags;
     let numToDisplay = props.numOfDisplay || 3;
 
-    const [anchorEl, setAnchorEl] = useState();
-
-    const handleClick = (event, tag) => {
-        console.log('tag clicked, display description', tag.description);
-        setAnchorEl(event.currentTarget);
-    }
-    const handleClose = () => {
-        setAnchorEl(null);
-    };
-
-    const open = Boolean(anchorEl);
-    const id = open ? 'simple-popover' : undefined;  
+ 
 
     return(
         <div>
             {
                 tagsArray.slice(0,numToDisplay).map((tag, index) => {
                     return(
-                            <Button key={tag.id} onClick={(event) => handleClick(event, tag)}>
-                                {tag.name}
-                                <Popover
-                                    id={id}
-                                    open={open}
-                                    onClose={handleClose}
-                                    anchorEl={anchorEl}
-                                    anchorOrigin={{
-                                    vertical: 'bottom',
-                                    horizontal: 'left',
-                                    }}
-                                >
-                                    <Typography sx={{ p: 2 }}>
-                                        {tag.description}
-                                    </Typography>
-                                </Popover>
-                            </Button>
+                        <Tag tag={tag} index={index} key={tag.id}/>
+                        // create new component and prop down. also click away
                     )
                 })
             }
